Auto stash before merge of "master" and "origin/master"

+++ GodVr/Assets/Scripts/Hello World/Interfaces/IDamagable.cs
@@ -1,4 +1,4 @@
 ﻿public interface IDamagable
 {
-    void Receive(int damage);
+    void Receive(ulong damage);
 }
